% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drivers_license}
\alias{drivers_license}
\title{Drivers license}
\format{A data set with nine variables: \code{id}, \code{age},
  \code{height}, \code{eye_color}, \code{hair_color}, \code{gender},
  \code{plate_number}, \code{car_make} and \code{car_model}.}
\usage{
drivers_license
}
\description{
Data for the drivers license details we have [SQL Murder](http://mystery.knightlab.com/walkthrough.html)
and [here] (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game [here](https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\examples{
# Load a bit of the dataset: drivers_license
head(drivers_license)
}
\keyword{datasets}
