% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facebook_event_checkin}
\alias{facebook_event_checkin}
\title{Facebook checkin event}
\format{A data set with four variables: \code{person_id}, \code{event_id},
  \code{event_name} and \code{date}.}
\usage{
facebook_event_checkin
}
\description{
Data for the facebook checkin events we have [SQL Murder](http://mystery.knightlab.com/walkthrough.html)
and [here] (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game [here](https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\examples{
# Load a bit of the dataset: facebook_event_checkin
head(facebook_event_checkin)
}
\keyword{datasets}
