% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{get_db}
\alias{get_db}
\title{A sample sqlite database}
\usage{
get_db()
}
\description{
This database is bundled with the package, and contains all data frames
in the datasets package.
}
\examples{
library(DBI)
db <- reclues::get_db()
dbListTables(db)

dbGetQuery(db, "SELECT * FROM crime_scene_report LIMIT 10")

dbDisconnect(db)
}
