% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Income}
\format{A data set with two variables: \code{ssn} and \code{annual_income}.}
\usage{
income
}
\description{
Data for the income information we have [SQL Murder](http://mystery.knightlab.com/walkthrough.html)
and [here] (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game [here](https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\examples{
# Load a bit of the dataset: income
head(income)
}
\keyword{datasets}
