% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interview}
\alias{interview}
\title{Interview}
\format{A data set with two variables: \code{person_id} and \code{transcript}.}
\usage{
interview
}
\description{
Data for the interviews conducted [SQL Murder](http://mystery.knightlab.com/walkthrough.html)
and [here] (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game [here](https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\examples{
# Load a bit of the dataset: interview
head(interview)
}
\keyword{datasets}
