% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{person}
\alias{person}
\title{Person}
\format{A data set with six variables: \code{id}, \code{name},
  \code{license_id}, \code{address_number}, \code{address_street_name} and \code{ssn}.}
\usage{
person
}
\description{
Data for the drivers license details we have [SQL Murder](http://mystery.knightlab.com/walkthrough.html)
and [here] (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game [here](https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\examples{
# Load a bit of the dataset: person
head(person)
}
\keyword{datasets}
