% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solution}
\alias{solution}
\title{Solution}
\format{A empty data set with two variables: \code{user} and \code{value}. The data should be inserted as
(1, "Name of criminal") once you have cracked the puzzle.}
\usage{
solution
}
\description{
This is where the answer to the mystery goes!
}
\keyword{datasets}
