% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_fit_now_check_in}
\alias{get_fit_now_check_in}
\title{Fit Now checkin details}
\format{A data set with four variables: \code{membership_id}, \code{check_in_date},
  \code{check_in_time} and \code{check_out_time}.}
\usage{
get_fit_now_check_in
}
\description{
Data for the Fit Now checkins as per the SQL Murder (http://mystery.knightlab.com/walkthrough.html)
and here (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game (https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\details{
\bold{How to play the game ....}

This package makes the data of the SQL Murder Mystery available in R. \cr\cr
There's been a Murder in SQL City!
The Murder Mystery is designed to be both a self-directed lesson to
learn R concepts and is also a fun game for experienced R users to solve an intriguing crime.
Please read the \bold{Readme file} for the details!
}
\examples{
# Load a bit of the dataset: get_fit_now_check_in
head(get_fit_now_check_in)
}
\keyword{datasets}
