% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_fit_now_member}
\alias{get_fit_now_member}
\title{Fit Now member details}
\format{A data set with five variables: \code{id}, \code{person_id},
  \code{name}, \code{membership_start_date} and \code{membership_status}.}
\usage{
get_fit_now_member
}
\description{
Data for the Fit Now members as per the SQL Murder (http://mystery.knightlab.com/walkthrough.html)
and here (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game (https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\details{
\bold{How to play the game ....}

This package makes the data of the SQL Murder Mystery available in R. \cr\cr
There's been a Murder in SQL City!
The Murder Mystery is designed to be both a self-directed lesson to
learn R concepts and is also a fun game for experienced R users to solve an intriguing crime.
Please read the \bold{Readme file} for the details!
}
\examples{
# Load a bit of the dataset: get_fit_now_member
head(get_fit_now_member)
}
\keyword{datasets}
