% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{person}
\alias{person}
\title{Person}
\format{A data set with six variables: \code{id}, \code{name},
  \code{license_id}, \code{address_number}, \code{address_street_name} and \code{ssn}.}
\usage{
person
}
\description{
Data for the person details as per the SQL Murder (http://mystery.knightlab.com/walkthrough.html)
and here (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version
of the SQL game (https://sql-murder-mystery.datasette.io/sql-murder-mystery).
}
\details{
\bold{How to play the game ....}

This package makes the data of the SQL Murder Mystery available in R. \cr\cr
There's been a Murder in SQL City!
The Murder Mystery is designed to be both a self-directed lesson to
learn R concepts and is also a fun game for experienced R users to solve an intriguing crime.
Please read the \bold{Readme file} for the details!
}
\examples{
# Load a bit of the dataset: person
head(person)
}
\keyword{datasets}
